Add distribution zip creation functionality
Implemented a new function to create a zip file of the distribution package, including all files from the 'dist' folder. Updated the print_summary function to display information about the created zip file, enhancing the build process and user guidance for distribution. This change streamlines the packaging of the application for users.